[mmp] Add new Profile subclasses for Xamarin Mac (full / system)
The fallback to MonoMacProfile is not correct if we ask for the
platform assembly - which blocks enabling the linker for that
assembly (only)
